Additional files (maven, json, testing etc.). I'm not familiar with most of them, I was said, these are for dynamic code etc. The cinema can now present a table statistics and different movies are now in different halls!

+++ src/main/java/com/egger/cinema/Ticket.java
@@ -1,8 +1,6 @@
 
 package com.egger.cinema;
 
-import java.math.BigDecimal;
-
-public record Ticket(int rowNr, int seatNr, BigDecimal price){
+public record Ticket(int rowNr, int seatNr, double price){
 
 }
